+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - [fix:remove the secondary report.](https://github.com/Tencent/spring-cloud-tencent/pull/899)
 - [fix:optimize instance circuit beaker.](https://github.com/Tencent/spring-cloud-tencent/pull/909)
 - [fix:optimize multi service registration and discovery.](https://github.com/Tencent/spring-cloud-tencent/pull/912)
+- [feature: improve circuit breaker usage.](https://github.com/Tencent/spring-cloud-tencent/pull/913)